pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "openhtf"
version = "1.6.1"
authors = [
  { name="The OpenHTF Authors"},
]
description = "OpenHTF, the open hardware testing framework."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "attrs>=19.3.0",
    "colorama>=0.4.6",
    "contextlib2>=21.6.0",
    "inflection>=0.5.1",
    "google-auth>=1.34.0",
    "mutablerecords>=0.4.1",
    "PyYAML>=6.0.2",
    "pyOpenSSL>=17.1.0",
    "requests>=2.27.1",
    "sockjs_tornado>=1.0.7",
    "tornado>=6,<6.5",
    "typing_extensions>=4.12.2",
]
license = {file = "LICENSE"}

[project.optional-dependencies]
usb_plugs = [
    "libusb1>=3.1.0",
    "M2Crypto>=0.42.0",
]
update_units = ["xlrd>=1.0.0"]
serial_collection_plug = ["pyserial>=3.5"]
examples = ["pandas>=2.2.3"]

[project.urls]
Homepage = "https://github.com/google/openhtf"

[tool.setuptools.packages]
find = {}

[tool.setuptools.package-data]
openhtf = [
    "output/web_gui/dist/*.*",
    "output/web_gui/dist/css/*",
    "output/web_gui/dist/js/*",
    "output/web_gui/dist/img/*",
    "output/web_gui/*.*",
]
